requirements.txt
django
djangorestframework
django-jsonfield
